rollout gen2: extract shared helpers to rollout_utils_gen2.py
_inject_flat_schema, _inject_tracer_inds, and _sample_to_batch were
duplicated verbatim in both rollout scripts. Move them to a shared module
and import from there.
