+++ credit/applications/rollout_to_netcdf_gen2.py
@@ -44,76 +44,13 @@ from credit.output import load_metadata, make_xarray, save_netcdf_increment
 from credit.postblock.gen1 import GlobalMassFixer, GlobalWaterFixer, GlobalEnergyFixer
 from credit.forecast import load_forecasts
 from credit.pbs import launch_script, launch_script_mpi
+from credit.applications.rollout_utils_gen2 import _inject_flat_schema, _inject_tracer_inds, _sample_to_batch
 
 logger = logging.getLogger(__name__)
 warnings.filterwarnings("ignore")
 os.environ["TF_CPP_MIN_LOG_LEVEL"] = "3"
 os.environ["OMP_NUM_THREADS"] = "1"
 os.environ["MKL_NUM_THREADS"] = "1"
-
-
-# ---------------------------------------------------------------------------
-# Config helpers
-# ---------------------------------------------------------------------------
-
-
-def _inject_flat_schema(conf):
-    """Inject v1-style flat keys into conf['data'] so output.py utilities work."""
-    if "variables" in conf["data"]:
-        return
-    src = conf["data"]["source"]["ERA5"]
-    v = src["variables"]
-    prog = v.get("prognostic") or {}
-    diag = v.get("diagnostic") or {}
-    conf["data"]["variables"] = prog.get("vars_3D", [])
-    conf["data"]["surface_variables"] = prog.get("vars_2D", [])
-    # diagnostic_variables is the list of names shown in xarray output
-    conf["data"]["diagnostic_variables"] = diag.get("vars_2D", []) if diag else []
-    # level_ids: actual pressure/model-level values for xarray coordinate
-    conf["data"]["level_ids"] = src.get("levels", list(range(conf["model"]["levels"])))
-    # scaler_type needed by save_netcdf_increment
-    if "scaler_type" not in conf["data"]:
-        conf["data"]["scaler_type"] = "std_new"
-
-
-def _inject_tracer_inds(conf):
-    """Compute tracer_inds for TracerFixer from v2 variable layout (same as train_gen2.py)."""
-    tracer_conf = conf.get("model", {}).get("post_conf", {}).get("tracer_fixer", {})
-    if not tracer_conf.get("activate", False) or "tracer_inds" in tracer_conf:
-        return
-    src = conf["data"]["source"]["ERA5"]
-    n_levels = len(src.get("levels", []))
-    v = src["variables"]
-    vars_3d = (v.get("prognostic") or {}).get("vars_3D", [])
-    vars_2d = (v.get("prognostic") or {}).get("vars_2D", [])
-    diag_2d = (v.get("diagnostic") or {}).get("vars_2D", [])
-    output_vars = [vn for vn in vars_3d for _ in range(n_levels)] + vars_2d + diag_2d
-    tracer_names = tracer_conf.get("tracer_name", [])
-    tracer_thres_cfg = tracer_conf.get("tracer_thres", [])
-    thres_map = dict(zip(tracer_names, tracer_thres_cfg))
-    tracer_inds, tracer_thres = [], []
-    for i, vn in enumerate(output_vars):
-        if vn in thres_map:
-            tracer_inds.append(i)
-            tracer_thres.append(thres_map[vn])
-    conf["model"]["post_conf"]["tracer_fixer"]["tracer_inds"] = tracer_inds
-    conf["model"]["post_conf"]["tracer_fixer"]["tracer_thres"] = tracer_thres
-    conf["model"]["post_conf"]["tracer_fixer"]["denorm"] = False
-
-
-def _sample_to_batch(sample, source_name):
-    """Wrap a single LocalDataset sample (no batch dim) into preblock-compatible dict.
-
-    Produces the MultiSourceDataset format expected by ERA5Normalizer and ConcatToTensor:
-        {"input": {source_name: {var_key: tensor}}, "metadata": {source_name: {...}}}
-    """
-    meta = {
-        k: torch.tensor(v).unsqueeze(0) if isinstance(v, (int, float)) else v for k, v in sample["metadata"].items()
-    }
-    return {
-        "input": {source_name: {k: v.unsqueeze(0) for k, v in sample["input"].items()}},
-        "metadata": {source_name: meta},
-    }
 
 
 # ---------------------------------------------------------------------------
